backend/requirements.txt
celery[redis]
django-celery-beat
django-celery-results
psycopg2-binary
django-environ
djangorestframework
django-storages
django-cors-headers
drf_yasg
django-debug-toolbar
whitenoise>=6.5
